requirements.txt
jupyterlab
ipywidgets
pymupdf
Pillow
python-docx